Fixed a whole bunch of memory leaks, still more to go

+++ ComponentFramework1.0 Bullet/ComponentFramework/MathUtil.h
@@ -9,7 +9,7 @@ class MathUtil
 public:
 	inline static float DegToRads(float deg)
 	{
-		return deg * DEGREES_TO_RADIANS;
+		return float(deg * DEGREES_TO_RADIANS);
 	}
 };
